new script editor ux
