+++ src/ui/views/session/editor/EditorView.css
@@ -1,77 +1,23 @@
+.EditorView {
+  --status-bar-height: 48px;
+  pointer-events: none;
+  padding: calc(var(--status-bar-height) + var(--sp-xs)) var(--sp-md) var(--sp-md);
+  width: 100%;
+  height: 100%;
+}
+
 .EditorView__leftPanel,
 .EditorView__rightPanel {
-  --status-bar-height: 48px;
-  --controls-height: 58px;
-  --controls-top: calc(var(--controls-height) + var(--sp-xs));
-  position: absolute;
-  top: calc(var(--status-bar-height) + var(--sp-xs));
-  bottom: calc(var(--controls-top) + var(--sp-xs));
-  pointer-events: none;
-  display: flex;
-  flex-direction: column;
-  width: 100%;
   max-width: 320px;
+  width: 100vw;
+}
+
+.EditorView__leftPanel,
+.EditorView__rightPanel,
+.EditorView__centerPanel {
+  pointer-events: none;
 
   & > * {
     pointer-events: all;
-  }
-}
-.EditorView__leftPanel {
-  left: var(--sp-md);
-}
-
-.EditorView__rightPanel {
-  right: var(--sp-md);
-}
-
-.EditorView__code {
-  --editor-height: 48px;
-  --controls-height: 58px;
-  --controls-top: calc(var(--controls-height) + var(--sp-xs));
-  top: calc(var(--editor-height) + var(--sp-xs));
-  /* bottom: calc(var(--controls-top) + var(--sp-xs)); */
-
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  margin: auto;
-  opacity: 0.8;
-  width: 50vw;
-  height: 88.5vh;
-}
-
-.EditorView__buttons {
-  display: flex;
-  flex-direction: row;
-
-  & > * {
-    flex-grow: 1;
-  }
-}
-
-.EditorView__resetScriptDialog {
-  display: flex;
-  padding: var(--sp-xs) var(--sp-md);
-}
-
-.EditorView__resetScriptDialog_content {
-  text-align: center;
-  height: auto;
-}
-
-.EditorView__resetScriptDialog_content .Content__children {
-  & > * {
-    margin-bottom: var(--sp-xs);
-  }
-}
-
-.EditorView__resetScriptDialog_buttons {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-evenly;
-  gap: var(--sp-sm);
-
-  & > * {
-    flex-grow: 1;
   }
 }
