+++ src/ui/views/session/editor/ScriptEditor.css
@@ -1,0 +1,39 @@
+.ScriptEditor {
+  width: 100%;
+  height: 100%;
+  user-select: none;
+  position: relative;
+  background-color: var(--bg-surface);
+  border-radius: var(--br-xs);
+  overflow: hidden;
+  box-shadow: var(--bs-surface);
+}
+
+.ScriptEditor__header {
+  padding: 0 3px 0 0;
+}
+.ScriptEditor__headerToolbar {
+  gap: 3px;
+}
+
+.ScriptEditor__code {
+  position: relative;
+}
+
+.ScriptEditor__themeBtn {
+  position: absolute;
+  right: var(--sp-md);
+  bottom: var(--sp-md);
+  background-color: var(--bg-surface);
+  box-shadow: var(--bs-surface);
+  border-radius: 50%;
+  overflow: hidden;
+
+  & .IconButton {
+    width: 40px;
+    height: 40px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+}
